+++ src/components/Card/index.ts
@@ -1,3 +1,3 @@
-// Card from copy-ui @ 2025-10-31.
+// Card from copy-ui @ 2025-11-25.
 
 export { Card } from "./card";