+++ src/components/Card/CHANGELOG.md
@@ -1,3 +1,6 @@
 - 2025-10-21: Initial version.
 - 2025-10-31: Add the sub-component `Card.Footer`, update styles to make it
   looks better, etc.
+- 2025-11-25: Add the sub-component `Card.ContentInScrollArea`, and support the
+  `withBorder` prop for `Card.Header` and `Card.Footer`. Some styles are updated
+  to make it looks better.